requirements.txt
streamlit
pandas
python-dotenv
crewai
crewai-tools 
langchain-anthropic